refactor: Validate com injeção de dependencia
